+++ src/effect.rs
@@ -1,13 +1,10 @@
-use std::cell::RefCell;
-use std::rc::Rc;
 use std::time::Duration;
 
-use ratatui::buffer::{Buffer, Cell};
-use ratatui::layout;
-use ratatui::layout::{Margin, Position, Rect};
-use ratatui::style::Color;
-use crate::{CellIterator, EffectTimer};
+use crate::widget::EffectSpan;
 use crate::shader::Shader;
+use crate::{CellFilter, CellIterator, EffectTimer};
+use ratatui::buffer::Buffer;
+use ratatui::layout::Rect;
 
 /// Represents an effect that can be applied to terminal cells.
 /// The `Effect` struct wraps a shader, allowing it to be configured
@@ -86,131 +83,6 @@ impl Effect {
     }
 }
 
-/// A filter mode enables effects to operate on specific cells.
-#[derive(Clone, Default)]
-pub enum CellFilter {
-    /// Selects every cell
-    #[default]
-    All,
-    /// Selects cells with matching foreground color
-    FgColor(Color),
-    /// Selects cells with matching background color
-    BgColor(Color),
-    /// Selects cells within the inner margin of the area
-    Inner(Margin),
-    /// Selects cells outside the inner margin of the area
-    Outer(Margin),
-    /// Selects cells with text
-    Text,
-    /// Selects cells that match all the given filters
-    AllOf(Vec<CellFilter>),
-    /// Selects cells that match any of the given filters
-    AnyOf(Vec<CellFilter>),
-    /// Selects cells that do not match any of the given filters
-    NoneOf(Vec<CellFilter>),
-    /// Negates the given filter
-    Not(Box<CellFilter>),
-    /// Selects cells within the specified layout, denoted by the index
-    Layout(layout::Layout, u16),
-    /// Selects cells by predicate function
-    PositionFn(Rc<RefCell<dyn Fn(Position) -> bool>>),
-}
-
-impl CellFilter {
-    pub fn position_fn<F>(f: F) -> Self
-        where F: Fn(Position) -> bool + 'static
-    {
-        CellFilter::PositionFn(Rc::new(RefCell::new(f)))
-    }
-}
-
-pub struct CellSelector {
-    inner_area: Rect,
-    strategy: CellFilter,
-}
-
-impl CellSelector {
-    fn new(area: Rect, strategy: CellFilter) -> Self {
-        let inner_area = Self::resolve_area(area, &strategy);
-
-        Self { inner_area, strategy }
-    }
-
-    fn resolve_area(area: Rect, mode: &CellFilter) -> Rect {
-        match mode {
-            CellFilter::All                  => area,
-            CellFilter::Inner(margin)        => area.inner(*margin),
-            CellFilter::Outer(margin)        => area.inner(*margin),
-            CellFilter::Text                 => area,
-            CellFilter::AllOf(_)             => area,
-            CellFilter::AnyOf(_)             => area,
-            CellFilter::NoneOf(_)            => area,
-            CellFilter::Not(m)               => Self::resolve_area(area, m.as_ref()),
-            CellFilter::FgColor(_)           => area,
-            CellFilter::BgColor(_)           => area,
-            CellFilter::Layout(layout, idx)  => layout.split(area)[*idx as usize],
-            CellFilter::PositionFn(_)        => area,
-        }
-    }
-
-    pub fn is_valid(&self, pos: Position, cell: &Cell) -> bool {
-        let mode = &self.strategy;
-
-        self.valid_position(pos, mode)
-            && self.is_valid_cell(cell, mode)
-    }
-
-    fn valid_position(&self, pos: Position, mode: &CellFilter) -> bool {
-        match mode {
-            CellFilter::All           => self.inner_area.contains(pos),
-            CellFilter::Layout(_, _)  => self.inner_area.contains(pos),
-            CellFilter::Inner(_)      => self.inner_area.contains(pos),
-            CellFilter::Outer(_)      => !self.inner_area.contains(pos),
-            CellFilter::Text          => self.inner_area.contains(pos),
-            CellFilter::AllOf(s)      => s.iter()
-                .all(|mode| mode.selector(self.inner_area).valid_position(pos, mode)),
-            CellFilter::AnyOf(s)      => s.iter()
-                .any(|mode| mode.selector(self.inner_area).valid_position(pos, mode)),
-            CellFilter::NoneOf(s)     => s.iter()
-                .all(|mode| !mode.selector(self.inner_area).valid_position(pos, mode)),
-            CellFilter::Not(m)        => self.valid_position(pos, m.as_ref()),
-            CellFilter::FgColor(_)    => self.inner_area.contains(pos),
-            CellFilter::BgColor(_)    => self.inner_area.contains(pos),
-            CellFilter::PositionFn(f) => f.borrow()(pos),
-        }
-    }
-
-    fn is_valid_cell(&self, cell: &Cell, mode: &CellFilter) -> bool {
-        match mode {
-            CellFilter::Text => {
-                if cell.symbol().len() == 1 {
-                    let ch = cell.symbol().chars().next().unwrap();
-                    ch.is_alphabetic() || ch.is_numeric() || ch == ' ' || "?!.,:;".contains(ch)
-                } else {
-                    false
-                }
-            },
-
-            CellFilter::AllOf(s) => {
-                s.iter()
-                    .all(|s| s.selector(self.inner_area).is_valid_cell(cell, s))
-            },
-
-            CellFilter::FgColor(color) => cell.fg == *color,
-            CellFilter::BgColor(color) => cell.bg == *color,
-
-            CellFilter::Not(m) => !self.is_valid_cell(cell, m.as_ref()),
-
-            _ => true,
-        }
-    }
-}
-
-impl CellFilter {
-    pub fn selector(&self, area: Rect) -> CellSelector {
-        CellSelector::new(area, self.clone())
-    }
-}
 
 impl Clone for Effect {
     fn clone(&self) -> Self {
@@ -219,6 +91,10 @@ impl Clone for Effect {
 }
 
 impl Shader for Effect {
+    fn name(&self) -> &'static str {
+        self.shader.name()
+    }
+
     fn process(&mut self, duration: Duration, buf: &mut Buffer, area: Rect) -> Option<Duration> {
         let area = self.shader.area().unwrap_or(area);
         self.shader.process(duration, buf, area)
@@ -252,6 +128,10 @@ impl Shader for Effect {
         self.shader.reverse()
     }
 
+    fn timer(&self) -> Option<EffectTimer> {
+        self.shader.timer()
+    }
+
     fn timer_mut(&mut self) -> Option<&mut EffectTimer> {
         self.shader.timer_mut()
     }
@@ -262,6 +142,13 @@ impl Shader for Effect {
 
     fn reset(&mut self) {
         self.shader.reset()
+    }
+
+    fn as_effect_span(&self, offset: Duration) -> EffectSpan
+    where
+        Self: Sized + Clone,
+    {
+        self.shader.as_ref().as_effect_span(offset)
     }
 }